# docs/swagger.yaml
basePath: /
definitions:
...
  orders.productItem:
    properties:
      available:
        type: boolean
      created_at:
        type: string
      id:
        type: integer
      price:
        type: string
      product_id:
        type: integer
      quantity:
        type: string
      request_id:
        type: integer
      total_sum:
        type: string
      unit:
        type: string
    type: object
  orders.requestDTO:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_available:
        type: boolean
      raw_amount:
        type: string
      raw_name:
        type: string
    type: object
  orders.requestInp:
    properties:
      raw_amount:
        type: string
      raw_name:
        type: string
    type: object
  orders.requestItem:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_available:
        type: boolean
      raw_amount:
        type: string
      raw_name:
        type: string
    type: object
...
  orders.showOrderResp:
    properties:
      comment:
        type: string
      created_at:
        type: string
      details_type:
        description: '"requests" | "products"'
        type: string
      id:
        type: integer
      location_id:
        type: integer
      products:
        items:
          $ref: '#/definitions/orders.productItem'
        type: array
      requests:
        items:
          $ref: '#/definitions/orders.requestItem'
        type: array
      status_id:
        type: integer
      total_sum:
        type: string
      user_id:
        type: integer
    type: object
  orders.updateReq:
    properties:
      comment:
        type: string
      location_id:
        type: integer
      requests:
        items:
          $ref: '#/definitions/orders.requestInp'
        type: array
    type: object
  orders.updateResp:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      location_id:
        type: integer
      requests:
        items:
          $ref: '#/definitions/orders.requestDTO'
        type: array
      status_id:
        type: integer
      total_sum:
        type: string
      user_id:
        type: integer
    type: object
  userlocations.bindUserToLocationReq:
    properties:
      location_id:
        type: integer
      role_id:
        description: null => без роли / стереть
        type: integer
      user_id:
        type: integer
    type: object
  userlocations.myLocationItem:
    properties:
      address:
        type: string
      granted_at:
        type: string
      location_id:
        type: integer
      name:
        type: string
      role_id:
        type: integer
      role_name:
        type: string
    type: object
...
info:
  contact: {}
  description: Horeka backend
  title: Horeka API
  version: "1.0"
paths:
...
  /orders/{id}:
    get:
      description: |-
        Возвращает один заказ текущего авторизованного пользователя.

        Заказ ищется только среди заказов пользователя (id + user_id). Чужие заказы не отдаются.
        Если заказ не найден — возвращается 404.

        Состав деталей зависит от статуса заказа:
        - если status_id = 1 (новый) — возвращаются заявки из order_requests, а поле details_type = "requests"
        - иначе — возвращаются товары из order_products, а поле details_type = "products"

        В ответе всегда присутствуют основные поля заказа (id, location_id, status_id, total_sum, comment, created_at)
        и один из списков: requests или products.
      parameters:
      - description: ID заказа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orders.showOrderResp'
      security:
      - BearerAuth: []
      summary: Получить детали заказа по ID
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: |-
        Обновляет заказ текущего авторизованного пользователя по ID.

        Редактирование разрешено только для заказов в статусе "новый" (status_id = 1).
        Если статус уже изменился (в обработке/закрыт/отменен и т.п.) — возвращается конфликт и заказ не меняется.

        Обновляются поля location_id и comment (comment необязателен: пустая строка будет сохранена как NULL).

        Список заявок (requests) перезаписывается полностью:
        старые записи из order_requests удаляются и создаются новые на основе переданного массива.
        Это сделано, чтобы редактирование было простым и однозначным: итоговый список = то, что пришло в запросе.

        В ответе возвращается обновленный заказ и новый список заявок с их ID и временем создания.
      parameters:
      - description: ID заказа
        in: path
        name: id
        required: true
        type: integer
      - description: 'Новые данные заказа: location_id, comment (опц.), requests'
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/orders.updateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orders.updateResp'
      security:
      - BearerAuth: []
      summary: Редактировать заказ (только пока он новый)
      tags:
      - orders
...
  /user-locations/bind:
    post:
      consumes:
      - application/json
      description: |-
        Создает или обновляет привязку пользователя к локации.

        Если привязки user_id + location_id еще нет — она будет создана.
        Если привязка уже существует — будет обновлено только поле role_id.

        role_id — необязательное поле.
        - role_id = null: привязка остается, роль будет очищена (role_id станет NULL)
        - role_id > 0: роль будет установлена/заменена на указанную

        Метод проверяет существование пользователя, локации и (если передан) роли.
      parameters:
      - description: user_id, location_id, role_id (опц.)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/userlocations.bindUserToLocationReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Привязать пользователя к локации (назначить роль)
      tags:
      - user_locations
  /user-locations/me:
    get:
      description: |-
        Возвращает все локации, к которым текущий пользователь привязан (таблица user_locations).

        В ответе для каждой локации возвращаются:
        - location_id, name, address
        - role_id и role_name (могут быть null, если роль не назначена)
        - granted_at — дата/время, когда была выдана привязка к локации

        Список отсортирован по granted_at по убыванию (сначала самые последние привязки).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/userlocations.myLocationItem'
            type: array
      security:
      - BearerAuth: []
      summary: Получить список моих локаций
      tags:
      - user_locations
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"